feat(phase5): Phase 5.3 - Unify compare() and fix() to delegate to canonical modules _compare.py and _fix.py
