+++ python/zedda/__init__.py
@@ -1024,305 +1024,8 @@ def _print_plain(p: Any) -> None:
 # ─────────────────────────────────────────────────────────────────
 #  compare() — diff two datasets for drift detection
 # ─────────────────────────────────────────────────────────────────
-def compare(
-    path_a, path_b, sample_size: int | None = None, correlate: bool = False
-) -> None:
-    """
-    Compare two datasets side by side for drift detection.
+from ._compare import compare
 
-    Shows schema differences, null rate changes, distribution
-    shifts, new categories, and a final verdict.
-
-    Args:
-        path_a (str): Path to the first dataset (e.g. training data).
-        path_b (str): Path to the second dataset (e.g. production data).
-        sample_size (int, optional): Max rows to read per file.
-
-    Example::
-
-        import zedda as zd
-        zd.compare("train.csv", "test.csv")
-    """
-    res_a, temp_a = _resolve_input(path_a)
-    res_b, temp_b = _resolve_input(path_b)
-    try:
-        p_a = _scan_wrapper(res_a, sample_size=sample_size, correlate=correlate)
-        p_b = _scan_wrapper(res_b, sample_size=sample_size, correlate=correlate)
-
-        if not _RICH_AVAILABLE or _console is None:
-            raise ZeddaError(
-                "Rich is required for terminal output. Install with: pip install rich"
-            )
-
-        name_a = (
-            "<DataFrame A>"
-            if temp_a
-            else (
-                Path(path_a).name
-                if isinstance(path_a, (str, Path))
-                else "<DataFrame A>"
-            )
-        )
-        name_b = (
-            "<DataFrame B>"
-            if temp_b
-            else (
-                Path(path_b).name
-                if isinstance(path_b, (str, Path))
-                else "<DataFrame B>"
-            )
-        )
-
-        crit_sym = _safe_symbol("✗", "[X]")
-        warn_sym = _safe_symbol("⚠", "[!]")
-        check_sym = _safe_symbol("✓", "[OK]")
-        arrow_r = _safe_symbol("→", "->")
-
-        # ── Header ────────────────────────────────────────────────────
-        _console.print(
-            f"\n[bold blue]zedda[/bold blue] [dim]v{__version__}[/dim]  "
-            f"[dim]·  compare mode[/dim]\n"
-        )
-        _console.print(
-            f"  [bold]A[/bold] : [cyan]{name_a}[/cyan]"
-            f"     [dim]{p_a.num_rows:,} rows  ·  {p_a.num_cols} cols[/dim]"
-        )
-        _console.print(
-            f"  [bold]B[/bold] : [cyan]{name_b}[/cyan]"
-            f"     [dim]{p_b.num_rows:,} rows  ·  {p_b.num_cols} cols[/dim]"
-        )
-
-        cols_a = {c.name: c for c in p_a.columns}
-        cols_b = {c.name: c for c in p_b.columns}
-        all_cols = list(dict.fromkeys(list(cols_a) + list(cols_b)))
-
-        critical_errors = 0
-        warnings_count = 0
-
-        # ── Section 1: Schema ─────────────────────────────────────────
-        _console.print(f"\n{_section_header('Schema')}")
-
-        # Column count
-        if p_a.num_cols == p_b.num_cols:
-            _console.print(
-                f"  [green]{check_sym}[/green]  Column count   : "
-                f"{p_a.num_cols} / {p_b.num_cols} match"
-            )
-        else:
-            diff = abs(p_a.num_cols - p_b.num_cols)
-            _console.print(
-                f"  [yellow]{warn_sym}[/yellow]  Column count   : "
-                f"{p_a.num_cols} vs {p_b.num_cols}  "
-                f"[yellow]MISMATCH[/yellow] [dim](±{diff} col{'s' if diff != 1 else ''} — "
-                f"expected if target/label column is absent in B)[/dim]"
-            )
-            warnings_count += 1
-
-        # Type mismatches and missing columns
-        type_match_count = 0
-        type_total = 0
-        for name in all_cols:
-            ca = cols_a.get(name)
-            cb = cols_b.get(name)
-
-            if not cb:
-                _console.print(
-                    f"  [yellow]{warn_sym}[/yellow]  {rich_escape(name):<16}: "
-                    f"[yellow]MISSING in {name_b}[/yellow]"
-                    f"  [dim](expected if this is the target/label column)[/dim]"
-                )
-                warnings_count += 1
-            elif not ca:
-                _console.print(
-                    f"  [yellow]{warn_sym}[/yellow]  {rich_escape(name):<16}: "
-                    f"[yellow]MISSING in {name_a}[/yellow]"
-                    f"  [dim](new column in {name_b})[/dim]"
-                )
-                warnings_count += 1
-            else:
-                type_total += 1
-                if ca.type_str != cb.type_str:
-                    _console.print(
-                        f"  [red]{crit_sym}[/red]  {rich_escape(name):<16}: "
-                        f"{name_a}={ca.type_str}  {name_b}={cb.type_str}   "
-                        f"[red]TYPE MISMATCH[/red]"
-                    )
-                    critical_errors += 1
-                else:
-                    type_match_count += 1
-
-        if type_total > 0:
-            _console.print(
-                f"  [green]{check_sym}[/green]  Types          : "
-                f"{type_match_count} / {type_total} match"
-            )
-
-        # ── Section 2: Null Rates ─────────────────────────────────────
-        _console.print(f"\n{_section_header('Null Rates')}")
-
-        for name in all_cols:
-            ca = cols_a.get(name)
-            cb = cols_b.get(name)
-            if not ca or not cb:
-                continue
-
-            delta = cb.null_pct - ca.null_pct
-            if delta > 5:
-                _console.print(
-                    f"  [yellow]{warn_sym}[/yellow]  {rich_escape(name):<16}: "
-                    f"{ca.null_pct:.1f}%  {arrow_r}  {cb.null_pct:.1f}%   "
-                    f"[yellow]SPIKE (+{delta:.1f}%)[/yellow]"
-                )
-                warnings_count += 1
-            elif delta > 0.1:
-                _console.print(
-                    f"  [yellow]{warn_sym}[/yellow]  {rich_escape(name):<16}: "
-                    f"{ca.null_pct:.1f}%  {arrow_r}  {cb.null_pct:.1f}%   "
-                    f"[dim](+{delta:.1f}%) minor increase[/dim]"
-                )
-            elif delta < -0.1:
-                _console.print(
-                    f"  [green]{check_sym}[/green]  {rich_escape(name):<16}: "
-                    f"{ca.null_pct:.1f}%  {arrow_r}  {cb.null_pct:.1f}%   "
-                    f"[dim]({delta:.1f}%) minor decrease[/dim]"
-                )
-            else:
-                _console.print(
-                    f"  [green]{check_sym}[/green]  {rich_escape(name):<16}: "
-                    f"{ca.null_pct:.1f}%  {arrow_r}  {cb.null_pct:.1f}%    "
-                    f"[dim]stable[/dim]"
-                )
-
-        # ── Section 3: Distribution Shift ─────────────────────────────
-        _console.print(f"\n{_section_header('Distribution Shift')}")
-
-        shifts = _compute_distribution_shift(p_a.columns, p_b.columns)
-        for s in shifts:
-            name = s["col_name"]
-            is_int = next(c.type_str == "int" for c in p_a.columns if c.name == name)
-            mean_a_s = _format_num(s["mean_a"], is_int)
-            mean_b_s = _format_num(s["mean_b"], is_int)
-            shift_pct = s["shift_pct"]
-            psi = s.get("psi", 0.0)
-            ks_stat = s.get("ks_stat", 0.0)
-            wd = s.get("wasserstein", 0.0)
-
-            # Format metrics string
-            metrics_parts = []
-            if abs(shift_pct) >= 1.0:
-                metrics_parts.append(f"mean: {shift_pct:+.1f}%")
-            if psi > 0.01:
-                metrics_parts.append(f"PSI: {psi:.3f}")
-            if ks_stat > 0.01:
-                metrics_parts.append(f"KS: {ks_stat:.3f}")
-            if wd > 0.01:
-                metrics_parts.append(f"WD: {wd:.2f}")
-
-            metrics_str = f"({', '.join(metrics_parts)})" if metrics_parts else ""
-
-            if not s["is_stable"] and s["is_shift"]:
-                # DRIFT
-                _console.print(
-                    f"  [red]{crit_sym}[/red]  {rich_escape(name):<16}: "
-                    f"mean {mean_a_s} {arrow_r} {mean_b_s}  "
-                    f"[red]DRIFT {metrics_str}[/red]"
-                )
-                critical_errors += 1
-            elif not s["is_stable"]:
-                # SHIFT
-                _console.print(
-                    f"  [yellow]{warn_sym}[/yellow]  {rich_escape(name):<16}: "
-                    f"mean {mean_a_s} {arrow_r} {mean_b_s}  "
-                    f"[yellow]SHIFT {metrics_str}[/yellow]"
-                )
-                warnings_count += 1
-            else:
-                # STABLE
-                _console.print(
-                    f"  [green]{check_sym}[/green]  {rich_escape(name):<16}: "
-                    f"mean {mean_a_s} {arrow_r} {mean_b_s}   "
-                    f"[dim]stable {metrics_str}[/dim]"
-                )
-
-        # ── Section 4: Category Drift ─────────────────────────────────
-        cat_diffs = _compute_category_diff(p_a.columns, p_b.columns)
-        if cat_diffs:
-            _console.print(f"\n{_section_header('Category Drift')}")
-            for c in cat_diffs:
-                name = c["col_name"]
-                if not c.get("overflowed"):
-                    new_b = c.get("new_in_b", [])
-                    missing_b = c.get("missing_in_b", [])
-                    jaccard = c.get("jaccard", 1.0)
-                    if new_b:
-                        new_sample = ", ".join(f"'{v}'" for v in new_b[:3])
-                        if len(new_b) > 3:
-                            new_sample += f" (+{len(new_b) - 3} more)"
-                        _console.print(
-                            f"  [yellow]{warn_sym}[/yellow]  {rich_escape(name):<16}: "
-                            f"[yellow]{len(new_b)} unseen value{'s' if len(new_b) != 1 else ''} in B[/yellow] "
-                            f"({new_sample})  [dim]Jaccard={jaccard:.2f}[/dim]"
-                        )
-                        warnings_count += 1
-                    elif missing_b:
-                        _console.print(
-                            f"  [green]{check_sym}[/green]  {rich_escape(name):<16}: "
-                            f"[dim]subset of A ({c['unique_b']}/{c['unique_a']} categories, no unseen values)[/dim]"
-                        )
-                    else:
-                        _console.print(
-                            f"  [green]{check_sym}[/green]  {rich_escape(name):<16}: "
-                            f"[dim]category sets match exactly ({c['unique_a']} unique)[/dim]"
-                        )
-                else:
-                    diff_approx = c.get("diff_approx", 0)
-                    if diff_approx > 0:
-                        _console.print(
-                            f"  [yellow]{warn_sym}[/yellow]  {rich_escape(name):<16}: "
-                            f"[dim]high cardinality (~{c['unique_a']} in A vs ~{c['unique_b']} in B)[/dim]"
-                        )
-                    else:
-                        _console.print(
-                            f"  [green]{check_sym}[/green]  {rich_escape(name):<16}: "
-                            f"[dim]cardinality stable (~{c['unique_a']} unique)[/dim]"
-                        )
-
-        # ── Section 5: Verdict ────────────────────────────────────────
-        _console.print(f"\n{_section_header('Verdict')}")
-
-        if critical_errors > 0:
-            _console.print(
-                f"  [bold red]{crit_sym}  FAIL[/bold red]  —  "
-                f"{critical_errors} critical error{'s' if critical_errors != 1 else ''}"
-                + (
-                    f" · {warnings_count} warning{'s' if warnings_count != 1 else ''}"
-                    if warnings_count > 0
-                    else ""
-                )
-            )
-            _console.print("  Safe to train : [bold red]NO[/bold red]")
-        elif warnings_count > 0:
-            _console.print(
-                f"  [bold yellow]{warn_sym}  WARN[/bold yellow]  —  "
-                f"{warnings_count} warning{'s' if warnings_count != 1 else ''}"
-            )
-            _console.print(
-                "  Safe to train : [bold yellow]REVIEW[/bold yellow]"
-                "  [dim]— check flagged shifts before proceeding[/dim]"
-            )
-        else:
-            _console.print(
-                f"  [bold green]{check_sym}  PASS[/bold green]  —  no issues found"
-            )
-            _console.print("  Safe to train : [bold green]YES[/bold green]")
-
-        _console.print()
-
-    finally:
-        if temp_a:
-            _cleanup_temp(res_a)
-        if temp_b:
-            _cleanup_temp(res_b)
 
 
 # ─────────────────────────────────────────────────────────────────
@@ -1738,276 +1441,11 @@ def ml_ready(
 #  apply=True returns an actual cleaned DataFrame (not just code)
 #  All generated code uses repr() for column names (SEC-P01)
 # ─────────────────────────────────────────────────────────────────
-def fix(
-    path, apply: bool = False, sample_size: int | None = None, correlate: bool = False
-) -> Any:
-    """
-    Scan a dataset and generate copy-paste-ready pandas fix code.
+# ─────────────────────────────────────────────────────────────────
+#  fix() — Automated ML fix code generator
+# ─────────────────────────────────────────────────────────────────
+from ._fix import fix
 
-    Automatically detects the most common data quality problems and
-    prints grouped, actionable pandas snippets you can paste directly
-    into your data preparation notebook or script.
-
-    Issues detected and fixed:
-
-    * **Missing values** — numeric columns get ``fillna(median())``;
-      string columns get ``fillna(mode()[0])``.
-    * **Extreme outliers** — columns where max > 10x mean get a
-      clip-at-99th-percentile suggestion.
-    * **ID columns** — integer columns with >95% unique values are
-      flagged as likely row IDs and suggested for dropping.
-    * **High-cardinality strings** — string columns with >50 unique
-      values get a label-encoding suggestion.
-
-    The output is grouped by issue type and ends with a clean
-    **"Copy-Paste Block"** containing all fixes in one place.
-
-    Args:
-        path (str):
-            Path to a ``.csv``, ``.parquet``, or ``.arrow`` file.
-        apply (bool, default False):
-            If ``True``, actually apply all fixes and return a clean
-            pandas DataFrame instead of just printing code.
-        sample_size (int, optional):
-            Max rows to sample for profiling.
-            FIX P-M22: Added for API consistency with profile/scan/compare.
-
-    Returns:
-        None when ``apply=False`` (prints to terminal).
-        ``pandas.DataFrame`` when ``apply=True``.
-
-    Example::
-
-        import zedda as zd
-
-        # Print fix suggestions only
-        zd.fix("data.csv")
-
-        # Actually apply all fixes and get a clean DataFrame
-        clean_df = zd.fix("data.csv", apply=True)
-    """
-    resolved_path, is_in_memory = _resolve_input(path)
-    try:
-        if not _RICH_AVAILABLE or _console is None:
-            raise ZeddaError(
-                "Rich is required for terminal output. Install with: pip install rich"
-            )
-
-        # Run the C++ engine silently
-        p = _scan_wrapper(resolved_path, sample_size=sample_size, correlate=correlate)
-
-        # Each entry: (display_line, code_line)
-        # display_line = what we show in the grouped section
-        # code_line    = what goes in the final copy-paste block
-        null_fixes = []  # Missing value imputation fixes
-        outlier_fixes = []  # Extreme outlier transform fixes
-        id_col_fixes = []  # Useless ID column drop fixes
-        encoding_fixes = []  # High-cardinality string encoding fixes
-
-        for col in p.columns:
-            issues = _detect_column_issues(col, p)
-            if not issues:
-                continue
-
-            for issue in issues:
-                action = _get_fix_action(col, issue)
-
-                # Format the tuple (display_line, code_line)
-                arrow_r = _safe_symbol("→", "->")
-                display_line = (
-                    f"  [cyan]{action['display']}[/cyan]  "
-                    f"[dim]{arrow_r} {action['comment']} {arrow_r} {action['fix_action'].split('—')[0].strip(' .').lower()}[/dim]"
-                )
-                code_line = f"{action['fix_code']}  # {action['comment']}"
-
-                if (
-                    action["action_type"] == "drop"
-                    and issue["type"] == "high_nulls"
-                    or action["action_type"] == "impute"
-                ):
-                    null_fixes.append((display_line, code_line))
-                elif action["action_type"] == "clip":
-                    outlier_fixes.append((display_line, code_line))
-                elif action["action_type"] == "drop" and issue["type"] != "high_nulls":
-                    id_col_fixes.append((display_line, code_line))
-                elif action["action_type"] == "encode":
-                    encoding_fixes.append((display_line, code_line))
-
-        # Check if there is anything to fix
-        all_fixes = null_fixes + outlier_fixes + id_col_fixes + encoding_fixes
-        if not all_fixes:
-            _console.print(
-                Panel(
-                    "[green]No fixes needed![/green]  "
-                    "Your dataset looks clean and ML-ready.",
-                    title="[bold green]zd.fix() - All Clear[/bold green]",
-                    border_style="green",
-                    expand=False,
-                )
-            )
-            return None
-
-        # Print summary header
-        n_issues = len(all_fixes)
-        summary = (
-            f"[bold]{n_issues} issue{'s' if n_issues > 1 else ''} found[/bold] "
-            f"across [cyan]{p.num_cols}[/cyan] columns.\n"
-            f"[dim]Scroll down for the full copy-paste block.[/dim]"
-        )
-        file_name = (
-            "<DataFrame>"
-            if is_in_memory
-            else (Path(path).name if isinstance(path, (str, Path)) else "<DataFrame>")
-        )
-        _console.print(
-            Panel(
-                summary,
-                title=f"[bold yellow]zd.fix() - {file_name}[/bold yellow]",
-                border_style="yellow",
-                expand=False,
-            )
-        )
-
-        sq_icon = _safe_symbol("◼", "[*]")
-        # Print each category with a section header
-        if null_fixes:
-            _console.print(
-                f"\n[bold red]{sq_icon}  MISSING VALUES[/bold red]  "
-                "[dim](fills nulls with median / mode)[/dim]"
-            )
-            for display, _ in null_fixes:
-                _console.print(display)
-
-        if outlier_fixes:
-            _console.print(
-                f"\n[bold magenta]{sq_icon}  OUTLIERS[/bold magenta]  "
-                "[dim](log1p shrinks extreme right-skewed values)[/dim]"
-            )
-            for display, _ in outlier_fixes:
-                _console.print(display)
-
-        if id_col_fixes:
-            _console.print(
-                f"\n[bold blue]{sq_icon}  ID COLUMNS[/bold blue]  "
-                "[dim](high-uniqueness integers — useless for ML)[/dim]"
-            )
-            for display, _ in id_col_fixes:
-                _console.print(display)
-
-        if encoding_fixes:
-            _console.print(
-                f"\n[bold cyan]{sq_icon}  ENCODING[/bold cyan]  "
-                "[dim](high-cardinality strings → numeric codes)[/dim]"
-            )
-            for display, _ in encoding_fixes:
-                _console.print(display)
-
-        # Print the final copy-paste block
-        _console.print(
-            "\n[bold]Copy-Paste Block:[/bold]  "
-            "[dim](paste this into your notebook or script)[/dim]"
-        )
-
-        # Print imports only if they are actually needed by the generated code
-        needs_numpy = bool(outlier_fixes)  # np.log1p
-        needs_pandas = bool(encoding_fixes)  # pd.Categorical
-        if needs_numpy:
-            _console.print("[dim]import numpy as np[/dim]")
-        if needs_pandas:
-            _console.print("[dim]import pandas as pd[/dim]")
-
-        for _, code in all_fixes:
-            _console.print(f"  [cyan]{code}[/cyan]")
-        _console.print()
-
-        # apply=True: actually execute the fixes and return a DataFrame
-        if apply:
-            try:
-                import numpy as np
-                import pandas as pd
-            except ImportError:
-                _console.print(
-                    "[red]pandas / numpy not installed — cannot apply fixes.[/red]\n"
-                    "Run: pip install pandas numpy"
-                )
-                return None
-
-            # FIX M-16: Use Path.suffix for consistency with the rest of the codebase.
-            if is_in_memory:
-                if "polars" in getattr(type(path), "__module__", ""):
-                    df = path.to_pandas()
-                else:
-                    df = path.copy()
-            else:
-                assert isinstance(resolved_path, str)
-                if Path(resolved_path).suffix.lower() == ".csv":
-                    df = pd.read_csv(resolved_path)
-                else:
-                    df = pd.read_parquet(resolved_path)
-
-            # Apply null fixes
-            # FIX P-C3: Guard mode() against empty Series (all-null column).
-            for col in p.columns:
-                if col.null_pct > 1:
-                    if col.null_pct > 50 and col.type_str in ("str", "unknown"):
-                        df = df.drop(columns=[col.name], errors="ignore")
-                    elif col.type_str in ("int", "float"):
-                        coerced = pd.to_numeric(df[col.name], errors="coerce")
-                        df[col.name] = coerced.fillna(coerced.median())
-                    elif col.type_str in ("str", "unknown"):
-                        m = df[col.name].mode()
-                        if not m.empty:
-                            df[col.name] = df[col.name].fillna(m[0])
-
-            # Apply outlier fixes.
-            # FIX P-C2: Use the SAME clip-at-99th-percentile fix that the
-            # displayed copy-paste block shows. Previously this called
-            # np.log1p(...) which created a new _log column — diverging from
-            # the documented fix and changing the schema.
-            # FIX P-M27: Reuse the canonical _is_outlier_column() predicate
-            # so apply=True matches the displayed fix exactly.
-            for col in p.columns:
-                if _is_outlier_column(col) and col.name in df.columns:
-                    safe = _safe_col_name(col.name)
-                    upper = pd.to_numeric(df[col.name], errors="coerce").quantile(0.99)
-                    if pd.notna(upper):
-                        df[col.name] = (
-                            pd.to_numeric(df[col.name], errors="coerce")
-                            .clip(upper=upper)
-                            .infer_objects(copy=False)
-                        )
-
-            # Apply ID column drops
-            id_cols = [
-                col.name
-                for col in p.columns
-                if col.type_str == "int" and col.unique_pct > 95
-            ]
-            if id_cols:
-                df = df.drop(columns=id_cols, errors="ignore")
-
-            # Apply encoding fixes
-            for col in p.columns:
-                if col.type_str in ("str", "unknown") and col.unique_approx > 50:
-                    if col.name in df.columns:
-                        df[col.name] = pd.Categorical(df[col.name]).codes
-
-            _console.print(
-                Panel(
-                    f"[green]✔ Applied {n_issues} fix{'es' if n_issues > 1 else ''}.[/green]  "
-                    f"DataFrame shape: [cyan]{df.shape}[/cyan]",
-                    title="[bold green]zd.fix(apply=True) — Done[/bold green]",
-                    border_style="green",
-                    expand=False,
-                )
-            )
-            return df
-
-        return None
-
-    finally:
-        if is_in_memory:
-            _cleanup_temp(resolved_path)
 
 
 # ─────────────────────────────────────────────────────────────────
